submit for install.sh

+++ install.sh
@@ -3,26 +3,62 @@ CUR_DIR=`pwd`
 cd ~
 HOME_DIR=`pwd`/
 CAMP_DIR=${CUR_DIR#${HOME_DIR}}
+CLEAN=$1
 
-# remove current environment
-rm -fr ~/.vim
-rm -f ~/.vimrc
-rm -f ~/.bashrc
-rm -f ~/.bash_profile
-rm -f ~/.minttyrc
-rm -f ~/.inputrc
+if [ "${CLEAN}" = "clean" ]
+then
+    # remove current environment
+    rm -fr ~/.vim
+    rm -fr ~/.vimSession
+    rm -f ~/.vimrc
+    rm -f ~/.viminfo
+    rm -f ~/.bashrc
+    rm -f ~/.bash_profile
+    rm -f ~/.minttyrc
+    rm -f ~/.inputrc
+    rm -f ~/.astylerc
+    exit
+fi
 
 # prepare environment
-ln -s ${CAMP_DIR}/vimrc ~/.vimrc
-ln -s ${CAMP_DIR}/bashrc ~/.bashrc
-ln -s ${CAMP_DIR}/bash_profile ~/.bash_profile
-ln -s ${CAMP_DIR}/minttyrc ~/.minttyrc
-ln -s ${CAMP_DIR}/inputrc ~/.inputrc
+if [ ! -f ~/.vimrc ]; then
+    ln -s ${CAMP_DIR}/vimrc ~/.vimrc
+fi
 
-source ~/.bashrc
-source ~/.bash_profile
-source ~/.minttyrc
+if [ ! -f ~/.bashrc ]; then
+    ln -s ${CAMP_DIR}/bashrc ~/.bashrc
+fi
 
+if [ ! -f ~/.bash_profile ]; then
+    ln -s ${CAMP_DIR}/bash_profile ~/.bash_profile
+fi
+
+if [ ! -f ~/.minttyrc ]; then
+    ln -s ${CAMP_DIR}/minttyrc ~/.minttyrc
+fi
+
+if [ ! -f ~/.inputrc ]; then
+    ln -s ${CAMP_DIR}/inputrc ~/.inputrc
+fi
+
+if [ ! -f ~/.astylerc ]; then
+    ln -s ${CAMP_DIR}/astylerc ~/.astylerc
+fi
+
+# source environment config
+if [ -f ~/.bashrc ]; then
+    source ~/.bashrc
+fi
+
+if [ -f ~/.bash_profile ]; then
+    source ~/.bash_profile
+fi
+
+if [ -f ~/.minttyrc ]; then
+    source ~/.minttyrc
+fi
+
+# build vim-work environment
 mkdir -p ~/.vim
 cp -fr ${CAMP_DIR}/colors ~/.vim
 cp -fr ${CAMP_DIR}/syntax ~/.vim
@@ -46,4 +82,6 @@ cp -f ack-* /usr/bin/ack-grep
 chmod 777 /usr/bin/ack-grep
 
 # install bundle plugin
-git clone https://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
+if [ ! -d ~/.vim/bundle/vundle ]; then
+    git clone https://github.com/gmarik/vundle.git ~/.vim/bundle/vundle
+fi
